% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cBase}
\alias{cBase}
\title{Function used to combine baseline functions in one plot}
\usage{
cBase(..., legend.title, legend.labels)
}
\arguments{
\item{...}{\code{ggplot} objects created by plotting \code{reReg} objects.}

\item{legend.title}{an optional character string to specify the legend title.}

\item{name}{an optional character string to specify the legend labels.}
}
\description{
Combine different plots into one.
}
