% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcf.R
\name{mcf.formula-class}
\alias{mcf.formula-class}
\title{The \code{mcf.formula} class is imported from \code{reda}.}
\description{
The class \code{mcf.formula} represents a mcf formula.
See \code{reda} for details.
}
