% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcf.R
\name{mcf}
\alias{mcf}
\title{The \code{mcf} function is imported from \code{reda}.}
\description{
An S4 class generic function that returns the mean cumulative function (mcf) estiamtes.
See \code{?reda::mcf} for more details.
}
