% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.reReg}
\alias{plot.reReg}
\title{Plotting Estimated Baseline Cumulative Rate Functions}
\usage{
\method{plot}{reReg}(x, smooth = FALSE, return.grob = FALSE,
  control = list(), ...)
}
\arguments{
\item{x}{an object of class \code{reReg}, usually returned by the \code{reReg} function.}

\item{smooth}{an optional logical value whether loess smoothing will be applied.}

\item{return.grob}{an optional logical value.
If "TRUE", a \code{ggplot2} plot grob will be returned.}

\item{...}{for future methods.}
}
\description{
Plot the estimated baseline cumulative rate function and harzard function if available.
}
\examples{
data(readmission, package = "frailtypack")
fit <- reReg(reSurv(t.stop, id, event, death) ~ sex + chemo,
             data = subset(readmission, id < 50),
             method = "am.XCHWY", se = "resampling", B = 20)
plot(fit)
}
\seealso{
\code{\link{reReg}}
}
\keyword{plot.reReg}
