% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCMF}
\alias{plotCMF}
\title{Produce Cumulative Sample Mean Function Plot}
\usage{
plotCMF(formula, data, onePanel = FALSE, return.grob = FALSE,
  control = list(), ...)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a "~" operator, and the terms on the right.
The response must be a recurrent event survival object as returned by function \code{reSurv}.}

\item{data}{an optional data frame in which to interpret the variables occurring in the "formula".}

\item{onePanel}{an optinoal logical value. If "TRUE", only one graphical panel will be displayed.}

\item{return.grob}{an optional logical value. If "TRUE", a \code{ggplot2} plot grob will be returned.}

\item{control}{a list of control parameters.}

\item{...}{for future developments.}
}
\description{
Plot the cumulative sample mean function for an \code{reSurv} object, with more flexible options.
}
\details{
The argument \code{control} consists of options with argument defaults to a list with the following values:
\describe{
  \item{xlab}{customizable x-label, default value is "Time".}
  \item{ylab}{customizable y-label, default value is "Subject".}
  \item{title}{customizable title, default value is "Recurrent event plot".}
  \item{terminal.name}{customizable label for terminal event, default value is "Terminal event".}
  \item{recurrent.name}{customizable label for recurrent event, default value is "Recurrent event plot".}
  \item{recurrent.types}{customizable label for recurrent event type, default value is NULL.}
  \item{alpha}{controls the transparency of points.}
}
}
\examples{
data(readmission)
plotCMF(reSurv(t.stop, event, death, id) ~ 1, data = readmission)
}
\seealso{
\code{\link{reSurv}}, \code{\link{plot.reSurv}}
}
\keyword{plot.reSurv}
