% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCSM}
\alias{plotCSM}
\title{Produce Cumulative Sample Mean Function Plots}
\usage{
plotCSM(formula, data, onePanel = FALSE, return.grob = FALSE,
  adjrisk = TRUE, control = list(), ...)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a "~" operator, and the predictors on the right.
The response must be a recurrent event survival object as returned by function \code{reSurv}.}

\item{data}{an optional data frame in which to interpret the variables occurring in the "\code{formula}".}

\item{onePanel}{an optional logical value indicating whether cumulative sample means (CSM) will be plotted in the same panel.
This is useful when comparing CSM from different groups.}

\item{return.grob}{an optional logical value indicating whether a \code{ggplot2} plot grob will be returned.}

\item{adjrisk}{an optional logical value indicating whether risk set will be adjusted. See Details.}

\item{control}{a list of control parameters.}

\item{...}{for future developments.}
}
\description{
Plot the cumulative sample mean function (CSM) for an \code{reSurv} object, with more flexible options.
}
\details{
The argument \code{control} consists of options with argument defaults to a list with the following values:
\describe{
  \item{xlab}{customizable x-label, default value is "Time".}
  \item{ylab}{customizable y-label, default value is "Subject".}
  \item{title}{customizable title, default value is "Recurrent event plot".}
  \item{terminal.name}{customizable label for terminal event, default value is "Terminal event".}
  \item{recurrent.name}{customizable label for recurrent event, default value is "Recurrent event plot".}
  \item{recurrent.types}{customizable label for recurrent event type, default value is \code{NULL}.}
  \item{alpha}{between 0 and 1, controls the transparency of points.}
}
}
\examples{
data(readmission, package = "frailtypack")
plotCSM(reSurv(t.stop, id, event, death) ~ 1, data = readmission)
plotCSM(reSurv(t.stop, id, event, death) ~ sex, data = readmission)
plotCSM(reSurv(t.stop, id, event, death) ~ sex, data = readmission, onePanel = TRUE)
}
\seealso{
\code{\link{reSurv}}, \code{\link{plot.reSurv}}
}
\keyword{plot.reSurv}
