% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recur.R
\name{Recur-class}
\alias{Recur-class}
\title{The \code{Recur} class is imported from \code{reda}.}
\description{
The class \code{Recur} is an S4 that represents a formula response for
recurrent event data model. See \code{reda} for details.
}
