% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recur.R
\name{Recur-pipe}
\alias{Recur-pipe}
\alias{\%to\%}
\alias{\%2\%}
\title{The \code{\%to\%} function is imported from \code{reda}}
\description{
This pipe operator specifies the time segments or recurrent episodes by endpoints.
See \code{reda} for more details.
}
\examples{
Recur(2:6, id = c(1, 1, 1, 2, 2))
Recur(2:6, id = c(1, 1, 1, 2, 2))
Recur(1:5 \%2\% 2:6, id = c(1, 1, 1, 2, 2))
}
