% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recur.R
\name{Recur-to}
\alias{Recur-to}
\alias{\%to\%}
\alias{\%2\%}
\title{Recurrent Episodes}
\usage{
time1 \%to\% time2

time1 \%2\% time2
}
\arguments{
\item{time1}{The left end-points of the recurrent episodes.}

\item{time2}{The right end-points of the recurrent episodes.}
}
\value{
A list that consists of two elements named
    \code{"time1"} and \code{"time2"}.
}
\description{
Specify time segements or recurrent episodes by endpoints.
}
\details{
This function is intended to be used for specifying the argument \code{time}
in function \code{\link{Recur}}.
}
