% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recur.R
\name{Recur}
\alias{Recur}
\alias{is.Recur}
\title{The \code{Recur} function is imported from \code{reda}.}
\description{
See \code{?reda::Recur} for more details.

See \code{?reda::is.Recur} for more details.
}
\details{
\itemize{
\item \code{\link[reda]{is.Recur}}: imported from the \code{reda}
package.
}
}
