\name{reReg-package}
\alias{reReg-package}
%\alias{aftgee}
\docType{package}
\title{
  Recurrent event regression 
}
\description{
  Recurrent event data arise frequently in various fields such as biomedical sciences,
  public health, engineering, and social sciences. In many instances, the observation of
  the recurrent event process can be stopped by the occurrence of a correlated failure
  event, such as death or treatment failure. As such, we implement a
  collection of regression models for recurrent event process and
  failure time.     
}
\details{
  \tabular{ll}{
    Package: reReg
    Type: Package
    Version: 1.0-0
    Date: 2015-10-30
    License: GPL(>=3)
    LazyLoad: yes
  }
}
\author{
  Sy Han Chiou \email{schiou@hsph.harvard.edu}
}
\references{
  Wang, M.C., Qin, J., and Chiang, C.T. (2001). Analyzing Recurrent
  Event Data with Informative Censoring. \emph{Journal of the American
    Statistical Association} \bold{96}{455}: 1057--1065.
  
  Ghosh, D. and D.Y. Lin (2003). Semiparametric Analysis of Recurrent
  Events Data in the Presence of Dependent Censoring. \emph{Biometrics},
  \bold{59}: 877 -- 885.

  Huang, C.Y. and Wang, M.C. (2004). Joint Modeling and Estimation
  for Recurrent Event Processes and Failure Time Data. \emph{Journal
    of the American Statistical Association} \bold{99}(468), 1153--1165.
}
\keyword{ reReg }
