% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recur.R
\name{summary.Recur-class}
\alias{summary.Recur-class}
\title{The \code{summary.Recur} class is imported from \code{reda}.}
\description{
The class \code{summary.Recur} is an S4 that represents the summary of a \code{Recur} object.
See \code{reda} for details.
}
