% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.reReg}
\alias{plot.reReg}
\title{Plotting Estimated Baseline Cumulative Rate Functions}
\usage{
\method{plot}{reReg}(x, ...)
}
\arguments{
\item{x}{an object of class \code{reReg}, usually returned by the \code{reReg} function.}

\item{...}{for future methods.}
}
\description{
Plot the estimated baseline cumulative rate function and harzard function if available.
}
\examples{
data(readmission)
fit <- reReg(reSurv(t.stop, event, death, id) ~ sex + chemo,
             data = subset(readmission, id < 50),
             method = "am.XCHWY", se = "resampling", B = 20)
plot(fit)
}
\seealso{
\code{\link{reReg}}
}
\keyword{plot.reReg}
