% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotHaz}
\alias{plotHaz}
\title{Plot the Baseline Cumulative Hazard Function for the Terminal Time}
\usage{
plotHaz(
  x,
  newdata = NULL,
  frailty = NULL,
  showName = FALSE,
  type = c("unrestricted", "bounded", "scaled"),
  smooth = FALSE,
  control = list(),
  ...
)
}
\arguments{
\item{x}{an object of class \code{reReg}, returned by the \code{reReg} function.}

\item{newdata}{an optional data frame contains variables to include in the calculation
of the cumulative rate function.
If omitted, the baseline rate function will be plotted.}

\item{frailty}{an optional vector to specify the shared frailty for \code{newdata}.
If \code{newdata} is given and \code{frailty} is not specified, the}

\item{showName}{an optional logical value indicating whether to label the curves
when \code{newdata} is specified.}

\item{type}{a character string specifying the type of rate function to be plotted.
Options are "unrestricted", "scaled", "bounded". See \bold{Details}.}

\item{smooth}{an optional logical value indicating whether to add a smooth curve
obtained from a monotone increasing P-splines implemented in package \code{scam}.}

\item{control}{a list of control parameters.}

\item{...}{graphical parameters to be passed to methods.
These include \code{xlab}, \code{ylab}, \code{main}, and more. See \bold{Details}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the baseline cumulative hazard function for an \code{reReg} object.
The 95\% confidence interval on the baseline cumulative rate function
}
\details{
The argument \code{control} consists of options with argument
defaults to a list with the following values:
\describe{
  \item{xlab}{customizable x-label, default value is "Time".}
  \item{ylab}{customizable y-label, default value is empty.}
  \item{main}{customizable title, default value is "Baseline cumulative hazard function".}
}
These arguments can also be passed down without specifying a \code{control} list.
}
\examples{
data(simDat)
fm <- Recur(t.start \%to\% t.stop, id, event, status) ~ x1 + x2

fit <- reReg(fm, data = simDat, model = "cox|cox", B = 0)
## Plot both the baseline cumulative rate and hazard function
plot(fit)
## Plot baseline cumulative hazard function
plotHaz(fit)
plotHaz(fit, smooth = TRUE)
}
\seealso{
\code{\link{reReg}} \code{\link{plot.reReg}}
}
\keyword{Plots}
