% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reReg.R
\name{reReg}
\alias{reReg}
\title{Fits Semiparametric Regression Models for Recurrent Event Data}
\usage{
reReg(formula, data, B = 200, method = "cox.LWYY",
  se = c("resampling", "bootstrap", "NULL"), control = list())
}
\arguments{
\item{formula}{a formula object, with the response on the left of a "~" operator, and the predictors on the right.
The response must be a recurrent event survival object as returned by function \code{Recur}.}

\item{data}{an optional data frame in which to interpret the variables occurring in the \code{"formula"}.}

\item{B}{a numeric value specifies the number of resampling for variance estimation.
When \code{B = 0}, variance estimation will not be performed.}

\item{method}{a character string specifying the underlying model. See \bold{Details}.}

\item{se}{a character string specifying the method for standard error estimation. See \bold{Details}.}

\item{control}{a list of control parameters.}
}
\description{
Fits a general (joint) semiparametric regression model for the recurrent event data,
where the rate function of the underlying recurrent event process and
the hazard function of the terminal event can be specified as a Cox-type model,
an accelerated mean model, an accelerated rate model, or a generalized scale-change model.
See details for model specifications.
}
\examples{
fm <- Recur(Time, id, event, status) ~ x1 + x2

## Joint Accelerated Mean Model
set.seed(1)
dat <- simSC(80, c(-1, 1), c(-1, 1), type = "am")
(fit <- reReg(Recur(Time, id, event, status) ~ x1 + x2, 
              data = dat, method = "am", se = "resampling", B = 20))
summary(fit)

## Generalized Scale-Change Model
set.seed(1)
dat <- simSC(100, c(-1, 1), c(-1, 1), type = "sc")
(fit <- reReg(Recur(Time, id, event, status) ~ x1 + x2, 
              data = dat, method = "sc|.", se = "resampling", B = 20))
summary(fit)
}
\references{
Xu, G., Chiou, S.H.,Yan, J., Marr, K., and Huang, C.-Y. (2019). Generalized Scale-Change Models for Recurrent Event
processes under informative censoring. \emph{Statistica Sinica}: pre-print.

Xu, G., Chiou, S.H., Huang, C.-Y., Wang, M.-C. and Yan, J. (2017). Joint Scale-change Models for Recurrent Events and Failure Time.
\emph{Journal of the American Statistical Association}, \bold{112}(518): 796--805.

Lin, D., Wei, L., Yang, I. and Ying, Z. (2000). Semiparametric Regression for the Mean and Rate Functions of Recurrent Events.
\emph{Journal of the Royal Statistical Society: Series B (Methodological)}, \bold{62}: 711--730.

Wang, M.-C., Qin, J., and Chiang, C.-T. (2001). Analyzing Recurrent Event Data with Informative Censoring.
\emph{Journal of the American Statistical Association}, \bold{96}(455): 1057--1065.

Ghosh, D. and Lin, D.Y. (2003). Semiparametric Analysis of Recurrent Events Data in the Presence of Dependent Censoring.
\emph{Biometrics}, \bold{59}: 877--885.

Huang, C.-Y. and Wang, M.-C. (2004). Joint Modeling and Estimation for Recurrent Event Processes and Failure Time Data.
\emph{Journal of the American Statistical Association}, \bold{99}(468): 1153--1165.

Ghosh, D. and Lin, D.Y. (2002). Marginal Regression Models for Recurrent and Terminal Events. \emph{Statistica Sinica}: 663--688.
}
\seealso{
\code{\link{Recur}}, \code{\link{simSC}}
}
