% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simDat}
\alias{simDat}
\title{Simulated dataset for demonstration}
\format{
A data frame with 274 rows and 6 variables.
}
\usage{
data(simDat)
}
\description{
A simulated data frame with variables
\describe{
  \item{id}{subjects identification}
  \item{Time}{observation time}
  \item{status}{terminal event indicator; 1 if a terminal event is recorded}
  \item{event}{recurrent event indicator; 1 if a recurrent event is recorded}
  \item{x1}{baseline covariate generated from a standard uniform distribution}
  \item{x2}{baseline covariate generated from a standard uniform distribution (independent from \code{z1}}
}
}
\details{
The sample dataset is generated by \code{set.seed(1); simSC(30, c(-1, 1), c(-1, 1))}.
See \code{\link{simSC}} for instruction on simulating recurrent event data from
scale-change models.
}
